chore: A bit of progress
